[ADDED]: added custom confirmation box component, added seperate layout for cart page when cart is empty, changed notification component position

+++ src/app/(components)/cart-item.tsx
@@ -5,14 +5,22 @@ import Link from "next/link";
 import { useDispatch, useSelector } from "react-redux";
 import { removeFromCart, incrementQuantity } from "../lib/cart/cartSlice";
 import { showNotification } from "../lib/notifications/notificationSlice";
+import { showConfirm } from "../lib/confirmation/confirmationSlice";
 
 const CartItemRow: FC<{ item: CartItem, index: number }> = ({ item, index }) => {
   const dispatch = useDispatch();
   const [quantity, setQuantity] = useState(0)
 
   const handleRemoveCartItem = (item: CartItem) => {
-    dispatch(removeFromCart(item))
-    dispatch(showNotification({id: Date.now(), message: 'Item removed from cart', type: 'success'}))
+    dispatch(showConfirm({
+      show: true,
+      message: "Remove item from cart?",
+      onConfirm: () => {
+        dispatch(removeFromCart(item))
+        dispatch(showNotification({ id: Date.now(), message: 'Item removed from cart', type: 'success' }))
+      }
+    }))
+
   }
   return (
     <>
@@ -35,7 +43,7 @@ const CartItemRow: FC<{ item: CartItem, index: number }> = ({ item, index }) => 
               <path d="M432 256c0 17.7-14.3 32-32 32L48 288c-17.7 0-32-14.3-32-32s14.3-32 32-32l352 0c17.7 0 32 14.3 32 32z" />
             </svg>
           </button>
-          <input type="number" name="prod-quantity" id="" value={item.quantity} onChange={(e) => setQuantity(Number(e.target.value)) } className="text-center text-lg" />
+          <input type="number" name="prod-quantity" id="" value={item.quantity} onChange={(e) => setQuantity(Number(e.target.value))} className="text-center text-lg" />
           <button className="btn plus" onClick={() => dispatch(incrementQuantity(item))}>
             <svg xmlns="http://www.w3.org/2000/svg" height="16" width="14" viewBox="0 0 448 512">
               <path d="M256 80c0-17.7-14.3-32-32-32s-32 14.3-32 32V224H48c-17.7 0-32 14.3-32 32s14.3 32 32 32H192V432c0 17.7 14.3 32 32 32s32-14.3 32-32V288H400c17.7 0 32-14.3 32-32s-14.3-32-32-32H256V80z" />
